% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bscui_elements.R
\name{add_bscui_elements}
\alias{add_bscui_elements}
\title{Add SVG elements to the UI}
\usage{
add_bscui_elements(proxy)
}
\arguments{
\item{proxy}{a \code{\link{bscui_Proxy}} object}
}
\value{
the provided proxy object
}
\description{
Add SVG elements to the UI
}
