% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/click_bscui_element.R
\name{click_bscui_element}
\alias{click_bscui_element}
\title{Trigger a click event on a clickable element}
\usage{
click_bscui_element(bscui, element_id, dbl_click = FALSE)
}
\arguments{
\item{bscui}{a \code{\link{bscui_Proxy}} object}

\item{element_id}{element identifier on which the click will be triggered}

\item{dbl_click}{logical indicating the type of click
(default: FALSE => single click is triggered)}
}
\value{
the provided bscui object
}
\description{
Trigger a click event on a clickable element
}
