% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_bscui_to_image.R
\name{export_bscui_to_image}
\alias{export_bscui_to_image}
\title{Save a bscui widget to an image file}
\usage{
export_bscui_to_image(widget, file, delay = 0.2, quiet = TRUE)
}
\arguments{
\item{widget}{a \code{\link{bscui}} object}

\item{file}{name of output file. Should end with an image file type
(.png, .jpg, .jpeg, or .webp) or .pdf.}

\item{delay}{Time to wait before taking screenshot, in seconds.
Sometimes a longer delay is needed for all assets to display properly.}

\item{quiet}{If TRUE (default), status updates via console messages are
suppressed.}
}
\description{
Save a bscui widget to an image file
}
\seealso{
\code{\link[webshot2:webshot]{webshot2::webshot()}}
}
