% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_displayed_svg.R
\name{get_displayed_svg}
\alias{get_displayed_svg}
\title{Get the displayed SVG}
\usage{
get_displayed_svg(bscui)
}
\arguments{
\item{bscui}{a \code{\link{bscui_Proxy}} object}
}
\description{
Get the displayed SVG
}
