% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_bscui_elements.R
\name{order_bscui_elements}
\alias{order_bscui_elements}
\title{Move svg elements in ui}
\usage{
order_bscui_elements(proxy)
}
\arguments{
\item{proxy}{a \code{\link{bscui_Proxy}} object}
}
\value{
the provided proxy object
}
\description{
Move svg elements in ui
}
