% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_svg_lines.R
\name{read_svg_lines}
\alias{read_svg_lines}
\title{Read an svg file as a character scalar}
\usage{
read_svg_lines(file)
}
\arguments{
\item{file}{file path}
}
\value{
A character vector of length 1
}
\description{
Read an svg file as a character scalar
}
