% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_bscui_elements.R
\name{remove_bscui_elements}
\alias{remove_bscui_elements}
\title{Remove SVG elements from the UI}
\usage{
remove_bscui_elements(proxy)
}
\arguments{
\item{proxy}{a \code{\link{bscui_Proxy}} object}
}
\value{
the provided proxy object
}
\description{
Remove SVG elements from the UI
}
