% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_bscui.R
\name{save_bscui}
\alias{save_bscui}
\title{Save a bscui widget to an HTML file}
\usage{
save_bscui(
  widget,
  file,
  selfcontained = TRUE,
  libdir = NULL,
  background = "white",
  title = class(widget)[[1]],
  knitrOptions = list()
)
}
\arguments{
\item{widget}{Widget to save}

\item{file}{File to save HTML into}

\item{selfcontained}{Whether to save the HTML as a single self-contained file
(with external resources base64 encoded) or a file with external resources
placed in an adjacent directory.}

\item{libdir}{Directory to copy HTML dependencies into (defaults to
filename_files).}

\item{background}{Text string giving the html background color of the widget.
Defaults to white.}

\item{title}{Text to use as the title of the generated page.}

\item{knitrOptions}{A list of \pkg{knitr} chunk options.}
}
\description{
A shortcut to the \code{\link[htmlwidgets:saveWidget]{htmlwidgets::saveWidget()}} function.
}
