% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_selection.R
\name{select_elements}
\alias{select_elements}
\title{Add UI elements to the selection}
\usage{
select_elements(bscui, element_ids)
}
\arguments{
\item{bscui}{a \code{\link{bscui_Proxy}} object}

\item{element_ids}{element identifiers to add to the selection}
}
\description{
Add UI elements to the selection
}
