% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_bscui_attributes.R
\name{set_bscui_attributes}
\alias{set_bscui_attributes}
\title{Set attributes of elements of a bscui widget}
\usage{
set_bscui_attributes(
  widget,
  element_attributes,
  to_ignore = NULL,
  targeted_tags = widget$x$structure_shapes,
  append = FALSE
)
}
\arguments{
\item{widget}{a \code{\link{bscui}} object}

\item{element_attributes}{a data frame with an \strong{id} column
providing the element identifier and one column per attribute name.}

\item{to_ignore}{identifiers of elements to ignore:
if those elements are children of elements to update they won't be updated}

\item{targeted_tags}{targeted_tags affected tag names
(by default: structure_shapes of the scui object)}

\item{append}{if TRUE the value will be concatenate with the existing value}
}
\value{
The modified \code{\link{bscui}} object
}
\description{
Set attributes of elements of a bscui widget
}
