% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_bscui_element_styles.R
\name{set_bscui_element_styles}
\alias{set_bscui_element_styles}
\title{Set the style of bscui ui elements}
\usage{
set_bscui_element_styles(
  bscui,
  element_styles,
  to_ignore = NULL,
  targeted_tags = NULL
)
}
\arguments{
\item{bscui}{a \code{\link{bscui_Proxy}} object}

\item{element_styles}{a data frame with an "id" column and
column per style to apply}

\item{to_ignore}{of elements to ignore: if those elements are
children of elements to update they won't be updated}

\item{targeted_tags}{affected tag names. If NULL (default),
the structure_shapes of the \code{\link{bscui}} object}
}
\value{
the provided \code{\link{bscui_Proxy}} object
}
\description{
Set the style of bscui ui elements
}
