% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_bscui_options.R
\name{set_bscui_options}
\alias{set_bscui_options}
\title{Set options of bscui widget}
\usage{
set_bscui_options(
  widget,
  show_menu,
  menu_width,
  zoom_min,
  zoom_max,
  zoom_step,
  clip,
  default_png_scale,
  selection_color,
  selection_opacity,
  hover_color,
  hover_opacity,
  structure_shapes,
  dblclick_timeout,
  hover_timeout
)
}
\arguments{
\item{widget}{a \code{\link{bscui}} object}

\item{show_menu}{if TRUE (default) control menu will be available}

\item{menu_width}{css width value (default: "30px")}

\item{zoom_min}{smallest zoom value (default: 0.5)}

\item{zoom_max}{largest zoom value (default: 20)}

\item{zoom_step}{zooming step: the larger the faster (default: 1.1)}

\item{clip}{if TRUE (default: FALSE), when the current zoom is 1, the viewBox
is automatically set to its original state (the drawing cannot be moved)}

\item{default_png_scale}{default value for scaling PNG export (default: 1)}

\item{selection_color}{color used to highlight selection (default: "orange")}

\item{selection_opacity}{opacity of selection highlight (default: 0.5)}

\item{hover_color}{a list of colors used to highlight hovered elements
(default: \code{list(button="yellow", selectable="cyan", none="transparent")})}

\item{hover_opacity}{opacity of hovered highlight (default: 0.5)}

\item{structure_shapes}{SVG shapes to considered as concrete
drawing
(default:
\code{c("rect", "circle", "ellipse", "line", "polyline", "polygon", "path")}:
"text" excluded)}

\item{dblclick_timeout}{minimum time in ms between 2 independant clicks
(default: 250)}

\item{hover_timeout}{time in ms before update hovered element (default: 100)}
}
\value{
The modified \code{\link{bscui}} object
}
\description{
Set options of bscui widget
}
