% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_bscui_selection.R
\name{set_bscui_selection}
\alias{set_bscui_selection}
\title{Pre-select UI elements in a bscui widget}
\usage{
set_bscui_selection(widget, selected)
}
\arguments{
\item{widget}{a \code{\link{bscui}} object}

\item{selected}{identifiers of pre-selected identifiers}
}
\value{
The modified \code{\link{bscui}} object
}
\description{
Pre-select UI elements in a bscui widget
}
