% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_bscui_ui_elements.R
\name{set_bscui_ui_elements}
\alias{set_bscui_ui_elements}
\title{Set UI elements of a bscui widget}
\usage{
set_bscui_ui_elements(widget, ui_elements)
}
\arguments{
\item{widget}{a \code{\link{bscui}} object}

\item{ui_elements}{NULL or a data frame with the following columns:
\itemize{
\item \strong{id}: the element identifier
\item \strong{ui_type}: either "selectable" (several elements can be selected),
"button" (action will be triggered on click), "none" (no ui)
\item \strong{title}: a description of the element to display on mouseover event
}}
}
\value{
The modified \code{\link{bscui}} object
}
\description{
Set UI elements of a bscui widget
}
