% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_bscui_selection.R
\name{update_bscui_selection}
\alias{update_bscui_selection}
\title{Replace current selection with given element identifiers}
\usage{
update_bscui_selection(proxy, element_ids)
}
\arguments{
\item{proxy}{a \code{\link{bscui_Proxy}} object}

\item{element_ids}{element identifiers to add to the selection; empty clear
the selection}
}
\value{
the provided proxy object
}
\description{
Replace current selection with given element identifiers
}
\examples{
if(interactive()){
   from_shiny <- new.env()
   shiny::runApp(system.file(
      "examples", "shiny-anatomogram", package = "bscui"
   ))
   for(n in names(from_shiny)){
      bscui(from_shiny[[n]]) |> print()
   }
}
}
