% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bscui.R
\name{bscui-shiny}
\alias{bscui-shiny}
\alias{bscuiOutput}
\alias{renderBscui}
\alias{bscuiProxy}
\alias{bscui_Proxy}
\title{Shiny bindings for bscui}
\usage{
bscuiOutput(outputId, width = "100\%", height = "400px")

renderBscui(expr, env = parent.frame(), quoted = FALSE)

bscuiProxy(shinyId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\\\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a bscui}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{shinyId}{single-element character vector indicating the shiny output ID
of the UI to modify}

\item{session}{the Shiny session object to which the UI belongs; usually the
default value will suffice}
}
\description{
Output and render functions for using bscui within Shiny applications.
}
\details{
The UI can be updated with \code{bscuiProxy}, using different methods:
\itemize{
\item \link{update_bscui_ui_elements}: change type and title of elements
\item \link{update_bscui_styles}: set style of UI elements
\item \link{update_bscui_attributes} set attributes of a UI element
\item \link{update_bscui_selection}: chose selected elements
\item \link{click_bscui_element}: trigger a single or double click on a UI element
\item \link{order_bscui_elements}: change elements order (e.g. move them forward)
\item \link{add_bscui_element}: add an SVG element to the UI
\item \link{remove_bscui_elements}: remove SVG elements from the UI
\item \link{get_bscui_svg}: get the displayed SVG in R session
}
}
\examples{
\dontrun{
   shiny::runApp(system.file(
      "examples", "shiny-anatomogram", package = "bscui"
   ))
}

}
