% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bscui.R
\name{bscui}
\alias{bscui}
\title{Build SVG/Shiny custom user interface}
\usage{
bscui(
  svg_txt,
  sanitize_attributes = TRUE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{svg_txt}{a character with svg code}

\item{sanitize_attributes}{logical indicating if
'<' and '>' characters in element attributes must be replaced by text}

\item{width, height}{widget width: must be a valid CSS unit (like \code{'100\\\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{hmtl identifier of the widget}
}
\value{
An \code{htmlwidget} object
}
\description{
Build SVG/Shiny custom user interface
}
