% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bscui.R
\name{bscui}
\alias{bscui}
\title{Build SVG/Shiny custom user interface}
\usage{
bscui(
  svg,
  ui_elements = NULL,
  show_menu = TRUE,
  menu_width = "20px",
  zoom_min = 0.5,
  zoom_max = 20,
  zoom_step = 1.1,
  clip = FALSE,
  default_png_scale = 1,
  selection_color = "orange",
  hover_color = list(button = "yellow", selectable = "grey"),
  dblclick_timeout = 250,
  hover_timeout = 500,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{svg}{an \code{xml_document} or a character with svg code}

\item{ui_elements}{NULL or a data frame with the following columns:
\itemize{
\item \strong{id}: the element identifier
\item \strong{ui_type}: either "selectable" (several elements can be selected),
"button" (action will be triggered on click), "none" (no ui)
\item \strong{title}: a description of the element to display on mouseover event
(support html tags). If NA, no description will be displayed.
\item \strong{visibility}: the css visibility of the element. NA values
are interpreted as "hidden".
\item \strong{opacity}: the opacity of the element between 0 and 1
(if NA, the value in the svg is kept)
\item \strong{fill}: the color used to fill the element
(if NA, the value in the svg is kept)
\item \strong{fill_opacity}: opacity of fill
(if NA, the value in the svg is kept)
\item \strong{stroke}: the color used for element border
(if NA, the value in the svg is kept)
\item \strong{stroke_width}: the width of the element border
(if NA, the value in the svg is kept)
\item \strong{stroke_opacity}: opacity of stroke
(if NA, the value in the svg is kept)
}}

\item{show_menu}{if TRUE (default) control menu will be available}

\item{menu_width}{css width value}

\item{zoom_min}{smallest zoom value}

\item{zoom_max}{largest zoom value}

\item{zoom_step}{zooming step: the larger the faster}

\item{clip}{if true, when the current zoom is 1, the viewBox is
automatically set to its original state (the drawing cannot be moved)}

\item{default_png_scale}{default value for scaling PNG export}

\item{selection_color}{color used to highlight selection}

\item{hover_color}{color used to highlight hovered element
(one for "button", one for "selectable", one for "none")}

\item{dblclick_timeout}{minimum time between 2 independant clicks}

\item{hover_timeout}{time before update hovered element}

\item{width, height}{widget width: must be a valid CSS unit (like \code{'100\\\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{hmtl identifier of the widget}
}
\description{
Build SVG/Shiny custom user interface
}
