% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/click_bscui_element.R
\name{click_bscui_element}
\alias{click_bscui_element}
\title{Trigger a click event on a clickable element}
\usage{
click_bscui_element(proxy, element_id, dbl_click = FALSE)
}
\arguments{
\item{proxy}{a \code{\link{bscui_Proxy}} object}

\item{element_id}{element identifier on which the click will be triggered}

\item{dbl_click}{logical indicating the type of click
(default: FALSE => single click is triggered)}
}
\value{
the provided proxy object
}
\description{
Trigger a click event on a clickable element
}
\examples{
if(interactive()){
   from_shiny <- new.env()
   shiny::runApp(system.file(
      "examples", "shiny-anatomogram", package = "bscui"
   ))
   for(n in names(from_shiny)){
      bscui(from_shiny[[n]]) |> print()
   }
}
}
