% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_bscui_to_image.R
\name{export_bscui_to_image}
\alias{export_bscui_to_image}
\title{Save a bscui widget to an image file}
\usage{
export_bscui_to_image(
  widget,
  file,
  selector = ".bscui",
  zoom = 1,
  quiet = TRUE,
  ...
)
}
\arguments{
\item{widget}{a \code{\link{bscui}} object}

\item{file}{name of output file. Should end with an image file type
(.png, .jpg, .jpeg, or .webp) or .pdf.}

\item{selector}{(\code{\link[webshot2:webshot]{webshot2::webshot()}}) one or more CSS selectors specifying
a DOM element to set the clipping rectangle to (default: ".bscui").}

\item{zoom}{(\code{\link[webshot2:webshot]{webshot2::webshot()}}) If TRUE (default),
status updates via console messages are suppressed.}

\item{quiet}{(\code{\link[webshot2:webshot]{webshot2::webshot()}}) a number specifying the zoom factor.}

\item{...}{additional parameters for \code{\link[webshot2:webshot]{webshot2::webshot()}}
suppressed.}
}
\description{
Save a bscui widget to an image file
}
\seealso{
\code{\link[webshot2:webshot]{webshot2::webshot()}}
}
