% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bscui_svg.R
\name{get_bscui_svg}
\alias{get_bscui_svg}
\title{Get the displayed SVG}
\usage{
get_bscui_svg(proxy)
}
\arguments{
\item{proxy}{a \code{\link{bscui_Proxy}} object}
}
\value{
the provided proxy object
}
\description{
Get the displayed SVG
}
\examples{
\dontrun{
   shiny::runApp(system.file(
      "examples", "shiny-anatomogram", package = "bscui"
   ))
}

}
