% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_bscui_elements.R
\name{remove_bscui_elements}
\alias{remove_bscui_elements}
\title{Remove SVG elements from the UI}
\usage{
remove_bscui_elements(proxy, element_ids)
}
\arguments{
\item{proxy}{a \code{\link{bscui_Proxy}} object}

\item{element_ids}{the identifiers of the elements to remove}
}
\value{
the provided proxy object
}
\description{
Remove SVG elements from the UI
}
\examples{
if(interactive()){
   from_shiny <- new.env()
   shiny::runApp(system.file(
      "examples", "shiny-anatomogram", package = "bscui"
   ))
   for(n in names(from_shiny)){
      bscui(from_shiny[[n]]) |> print()
   }
}
}
