% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_bscui_elements.R
\name{select_bscui_elements}
\alias{select_bscui_elements}
\title{Replace current selection with given element identifiers}
\usage{
select_bscui_elements(proxy, element_ids)
}
\arguments{
\item{proxy}{a \code{\link{bscui_Proxy}} object}

\item{element_ids}{element identifiers to add to the selection; empty clear
the selection}
}
\value{
the provided proxy object
}
\description{
Replace current selection with given element identifiers
}
\examples{
\dontrun{
   shiny::runApp(system.file(
      "examples", "shiny-anatomogram", package = "bscui"
   ))
}

}
